Conflicts:

	gui/Label.cpp
	screen/Screen.cpp
	screen/ScreenHelp.cpp
	screen/ScreenInfo.cpp
	screen/choose/ScreenChoose.cpp
	screen/choose/ScreenChooseGame.cpp
	screen/editor/ScreenEditorMain.cpp
	screen/game/ScreenGameEnd.cpp
	screen/game/ScreenGameMain.cpp
	screen/game/ScreenGameMain1.cpp
	screen/game/ScreenGameMain2.cpp
	screen/main/ScreenMain.cpp

+++ objects.cpp
@@ -77,7 +77,7 @@ void drawRingStrip(int corners, float progress, int texture)
 
 	float diff = p2 - p0;
 
-	int steps = (diff / step + 0.5f);
+	int steps = (int) (diff / step + 0.5f);
 
 	glColor3f(1.0f, 1.0f, 0.0f);
